リファクタリング 削除時にログアウトとメモのprivate化を実行
